+++ c3DSkit/graphics.c
@@ -1,19 +1,62 @@
 #include "graphics.h"
+#include <stdio.h>
 
 #define TESTNAME(name, value, id) if(strcmp(name,value)==0){return Py_BuildValue("i",id);}
 #define TESTPXSIZE(format, id, pxsize) if(format==id){return pxsize;}
 
+#define BC4_BYTESPERTEXEL 8
+
 enum TEXTURE_FORMATS{
 	RGBA8, RGB8, RGBA5551, RGB565, RGBA4,
-	LA8, HILO8, L8, A8, LA4, L4, A4,
-	ETC1, ETC1A4, BC1, BC2, BC3, BC4, BC5,
-	RGBA8_SRGB, BC1_SRGB, BC2_SRGB, BC3_SRGB
+	LA8, RG8, L8, A8, LA4, L4, A4,
+	ETC1, ETC1A4, BC1, BC2, BC3, BC4, BC5, BC6H, BC7,
+	RGBA8_SRGB, BC1_SRGB, BC2_SRGB, BC3_SRGB,
+	BC4_SNORM, BC5_SNORM, BC6H_SF16, BC7_SRGB,
 };
 
 uint8_t ETC1_MODIFIERS[8][2] = {
 	{2, 8}, {5, 17}, {9, 29}, {13, 42},
 	{18, 60}, {24, 80}, {33, 106}, {47, 183}
 };
+
+// Swizzling is taken from gdkchan's BnTxx
+
+typedef struct {
+	int bhshift;
+	int bppshift;
+	int bhmask;
+	int xshift;
+	int gobstride;
+} Swizzle;
+
+static int _getShift(int value){
+	if (value == 0) return 0;
+	int shift = 0;
+	while (((value >> shift) & 1) == 0){ shift++; }
+	return shift;
+}
+
+static void _makeSwizzle(Swizzle* swizzle, int width, int bpp, int blockheight){
+	swizzle->bhmask = blockheight * 8 - 1;
+	swizzle->bhshift = _getShift(blockheight * 8);
+	swizzle->bppshift = _getShift(bpp);
+	swizzle->gobstride = 512 * blockheight * (int)ceil(width * bpp / 64.0);
+	swizzle->xshift = _getShift(512 * blockheight);
+}
+
+static int _getSwizzleOffset(Swizzle* swizzle, int x, int y){
+	x <<= swizzle->bppshift;
+	int pos = (y >> swizzle->bhshift) * swizzle->gobstride;
+	pos += (x >> 6) << swizzle->xshift;
+	pos += ((y & swizzle->bhmask) >> 3) << 9;
+	pos += ((x & 0x3f) >> 5) << 8;
+	pos += ((y & 0x07) >> 1) << 6;
+	pos += ((x & 0x1f) >> 4) << 5;
+	pos += ((y & 0x01) >> 0) << 4;
+	pos += ((x & 0x0f) >> 0) << 0;
+	return pos;
+}
+	
 
 static int getPixelSize(int format){
 	TESTPXSIZE(format, RGBA8, 4);
@@ -22,7 +65,7 @@ static int getPixelSize(int format){
 	TESTPXSIZE(format, RGB565, 2);
 	TESTPXSIZE(format, RGBA4, 2);
 	TESTPXSIZE(format, LA8, 2);
-	TESTPXSIZE(format, HILO8, 2);
+	TESTPXSIZE(format, RG8, 2);
 	TESTPXSIZE(format, L8, 1);
 	TESTPXSIZE(format, A8, 1);
 	TESTPXSIZE(format, LA4, 1);
@@ -33,9 +76,9 @@ static void _extractTiledTexture(uint8_t* input, uint8_t* output, int width, int
 	int tilesx = (int)ceil((double)width / 8);
 	int tilesy = (int)ceil((double)height / 8);
 	int datawidth = 1 << (int)ceil(LOG2((double)width));
-	int dataheight = 1 << (int)ceil(LOG2((double)height));
+	//int dataheight = 1 << (int)ceil(LOG2((double)height));
 	int totalx = (int)ceil((double)datawidth / 8.0);
-	int totaly = (int)ceil((double)dataheight / 8.0);
+	//int totaly = (int)ceil((double)dataheight / 8.0);
 	int pxsize = getPixelSize(format);
 	for (int ytile = 0; ytile < tilesy; ytile++){
 		for (int xtile = 0; xtile < tilesx; xtile++){
@@ -94,8 +137,11 @@ static void _extractTiledTexture(uint8_t* input, uint8_t* output, int width, int
 											} else if (format == LA8){
 												r = g = b = input[inpos + 1];
 												a = input[inpos];
-											} else if (format == HILO8){
-												// ?
+											} else if (format == RG8){
+												r = input[inpos + 1];
+												g = input[inpos];
+												b = 0;
+												a = 0xFF;
 											} else if (format == L8){
 												r = g = b = input[inpos];
 												a = 0xFF;
@@ -135,8 +181,11 @@ static void _extractTiledTexture(uint8_t* input, uint8_t* output, int width, int
 											} else if (format == LA8){
 												r = g = b = input[inpos];
 												a = input[inpos + 1];
-											} else if (format == HILO8){
-												// ?
+											} else if (format == RG8){
+												r = input[inpos];
+												g = input[inpos + 1];
+												b = 0;
+												a = 0xFF;
 											} else if (format == L8){
 												r = g = b = input[inpos];
 												a = 0xFF;
@@ -162,6 +211,8 @@ static void _extractTiledTexture(uint8_t* input, uint8_t* output, int width, int
 		}
 	}
 }
+
+//ETC1 decompression is inspirated from ObsidianX's 3dstools BFLIM implementation
 
 static int ETC1DiffComplement(int val, int bits){
 	if ((val >> (bits - 1)) == 0){
@@ -256,50 +307,51 @@ static void _extractETC1Texture(uint8_t* input, uint8_t* output, int width, int 
 	}
 }
 
-static void _extractBC4Texture(uint8_t* input, uint8_t* output, int width, int height, int format){
-	int tilew = 1 << (int)ceil(LOG2(width / 8));
-	int tileh = 1 << (int)ceil(LOG2(height / 8));
-	uint8_t indices[16];
-	uint8_t values[8];
-	int inpos = 0;
-	for (int ytile = 0; ytile < tileh; ytile++){
-		for (int xtile = 0; xtile < tilew; xtile++){
-			for (int yblock = 0; yblock < 2; yblock++){
-				for (int xblock = 0; xblock < 2; xblock++){
-					values[0] = input[inpos++];
-					values[1] = input[inpos++];
-					for (int i = 0; i < 16; i += 8){
-						indices[i] = input[inpos + 2] & 0x07;
-						indices[i + 1] = (input[inpos + 2] >> 3) & 0x07;
-						indices[i + 2] = ((input[inpos + 1] & 1) << 2) | (input[inpos + 2] >> 6);
-						indices[i + 3] = (input[inpos + 1] >> 1) & 0x07;
-						indices[i + 4] = (input[inpos + 1] >> 4) & 0x07;
-						indices[i + 5] = (input[inpos + 1] >> 7) | ((input[inpos] & 0x03) << 1);
-						indices[i + 6] = (input[inpos] >> 2) & 0x07;
-						indices[i + 7] = (input[inpos] >> 5) & 0x07;
-						inpos += 3;
-					}
-					if (values[0] > values[1]){
-						for (int i = 2; i < 8; i++){
-							values[i] = ((8 - i) * values[0] + (i - 1) * values[1]) / 7;
-						}
+static void _extractBC4Texture(uint8_t* input, uint8_t* output, int width, int height, int format, int swizzlesize, bool littleendian){
+	int twidth = (width + 3) / 4;
+	int theight = (height + 3) / 4;
+	uint8_t lums[8];
+	Swizzle swizzle;
+	//int offset = 0;
+	_makeSwizzle(&swizzle, twidth, BC4_BYTESPERTEXEL, swizzlesize);
+	for (int ytile = 0; ytile < theight; ytile++){
+		for (int xtile = 0; xtile < twidth; xtile++){
+			int offset = _getSwizzleOffset(&swizzle, xtile, ytile);
+			lums[0] = input[offset];
+			lums[1] = input[offset + 1];
+			for (int i = 2; i < 8; i++){
+				if (lums[0] > lums[1]){
+					lums[i] = (uint8_t)(((8 - i) * lums[0] + (i - 1) * lums[1]) / 7);
+				} else if (i < 6){
+					lums[i] = (uint8_t)(((6 - i) * lums[0] + (i - 1) * lums[1]) / 7);
+				} else if (i == 6){
+					lums[i] = 0;
+				} else if (i == 7){
+					lums[i] = 0xFF;
+				}
+			}
+			int pixels1, pixels2;
+			if (littleendian){
+				pixels1 = input[offset + 2] | (input[offset + 3] << 8) | (input[offset + 4] << 16);
+				pixels2 = input[offset + 5] | (input[offset + 6] << 8) | (input[offset + 7] << 16);
+			} else {
+				pixels1 = input[offset + 4] | (input[offset + 3] << 8) | (input[offset + 2] << 16);
+				pixels2 = input[offset + 7] | (input[offset + 6] << 8) | (input[offset + 5] << 16);
+			}
+			int lum;
+			for (int ypix = 0; ypix < 4; ypix++){
+				for (int xpix = 0; xpix < 4; xpix++){
+					int ypos = ytile * 4 + ypix;
+					int xpos = xtile * 4 + xpix;
+					int outpos = (ypos * width + xpos) * 4;
+					int codeindex = ypix * 4 + xpix;
+					if (codeindex < 8){
+						lum = lums[(pixels1 >> (codeindex * 3)) & 7];
 					} else {
-						for (int i = 2; i < 6; i++){
-							values[i] = ((6 - i) * values[0] + (i - 1) * values[1]) / 7;
-						}
-						values[6] = 0;
-						values[7] = 0xFF;
+						lum = lums[(pixels2 >> ((codeindex - 8) * 3)) & 7];
 					}
-					for (int ypix = 0; ypix < 4; ypix++){
-						for (int xpix = 0; xpix < 4; xpix++){
-							int xpos = xtile * 8 + xblock * 4 + xpix;
-							int ypos = ytile * 8 + yblock * 4 + ypix;
-							int outpos = (ypos * width + xpos) * 4;
-							int index = ypix * 4 + xpix;
-							output[outpos] = output[outpos + 1] = output[outpos + 2] = values[indices[index]];
-							output[outpos + 3] = 0xFF;
-						}
-					}
+					output[outpos] = output[outpos + 1] = output[outpos + 2] = lum;
+					output[outpos + 3] = 0xFF;
 				}
 			}
 		}
@@ -309,11 +361,8 @@ static void _extractBC4Texture(uint8_t* input, uint8_t* output, int width, int h
 PyObject* extractTiledTexture(PyObject* self, PyObject* args){
 	PyArrayObject* input_obj;
 	PyArrayObject* output_obj;
-	int width;
-	int height;
-	int format;
-	int littleendian;
-	if (!PyArg_ParseTuple(args, "O!O!iiii", &PyArray_Type, &input_obj, &PyArray_Type, &output_obj, &width, &height, &format, &littleendian)){
+	int width, height, format, littleendian, swizzlesize;
+	if (!PyArg_ParseTuple(args, "O!O!iiiii", &PyArray_Type, &input_obj, &PyArray_Type, &output_obj, &width, &height, &format, &swizzlesize, &littleendian)){
 		return NULL;
 	}
 	uint8_t* input = (uint8_t*)PyArray_DATA(input_obj);
@@ -321,7 +370,7 @@ PyObject* extractTiledTexture(PyObject* self, PyObject* args){
 	if (format == ETC1 || format == ETC1A4){
 		_extractETC1Texture(input, output, width, height, format);
 	} else if (format == BC4){
-		_extractBC4Texture(input, output, width, height, format);
+		_extractBC4Texture(input, output, width, height, format, swizzlesize, (bool)littleendian);
 	} else {
 		_extractTiledTexture(input, output, width, height, format, (bool)littleendian);
 	}
@@ -339,7 +388,7 @@ PyObject* getTextureFormatId(PyObject* self, PyObject* args){
 	TESTNAME(name, "RGB565", RGB565);
 	TESTNAME(name, "RGBA4", RGBA4);
 	TESTNAME(name, "LA8", LA8);
-	TESTNAME(name, "HILO8", HILO8);
+	TESTNAME(name, "RG8", RG8);
 	TESTNAME(name, "L8", L8);
 	TESTNAME(name, "A8", A8);
 	TESTNAME(name, "LA4", LA4);
@@ -352,9 +401,15 @@ PyObject* getTextureFormatId(PyObject* self, PyObject* args){
 	//TESTNAME(name, "BC3", BC3);
 	TESTNAME(name, "BC4", BC4);
 	//TESTNAME(name, "BC5", BC5);
+	//TESTNAME(name, "BC6H", BC6H);
+	//TESTNAME(name, "BC7", BC7);
 	//TESTNAME(name, "RGBA8_SRGB", RGBA8_SRGB);
 	//TESTNAME(name, "BC1_SRGB", BC1_SRGB);
 	//TESTNAME(name, "BC2_SRGB", BC2_SRGB);
 	//TESTNAME(name, "BC3_SRGB", BC3_SRGB);
+	//TESTNAME(name, "BC4_SNORM", BC4_SNORM);
+	//TESTNAME(name, "BC5_SNORM", BC5_SNORM);
+	//TESTNAME(name, "BC6H_SF16", BC6H_SF16);
+	//TESTNAME(name, "BC7_SRGB", BC7_SRGB);
 	return Py_BuildValue("I", 0xFF);
 }